🔨refactor: Use Long in LogEntryIndex instead of Int

+++ src/main/scala/lerna/akka/entityreplication/raft/model/LogEntryIndex.scala
@@ -9,7 +9,7 @@ object LogEntryIndex {
   }
 }
 
-case class LogEntryIndex(underlying: Int) extends Ordered[LogEntryIndex] {
+case class LogEntryIndex(underlying: Long) extends Ordered[LogEntryIndex] {
   require(underlying >= 0)
 
   def next(): LogEntryIndex = copy(underlying + 1)
